Upgraded Ui, Logical flow and reoprts section upgraded to next level

+++ lib/features/customer/data/models/customer_model.dart
@@ -22,7 +22,7 @@ class CustomerModel extends Customer {
       id: map['id'],
       name: map['name'],
       phone: map['phone'],
-      points: map['points'] ?? 0,
+      points: (map['points'] ?? 0).toDouble(),
     );
   }
